+++ awscli/customizations/s3/executor.py
@@ -11,37 +11,49 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 import logging
-from six.moves import queue as Queue
+from six.moves import queue
 import sys
 import threading
 
-from awscli.customizations.s3.utils import NoBlockQueue, uni_print, \
-        IORequest, IOCloseRequest
+from awscli.customizations.s3.utils import uni_print, \
+        IORequest, IOCloseRequest, StablePriorityQueue
+from awscli.customizations.s3.tasks import OrderableTask
 
 
 LOGGER = logging.getLogger(__name__)
-QUEUE_END_SENTINEL = object()
+
+
+class ShutdownThreadRequest(OrderableTask):
+    PRIORITY = 11
+
+    def __init__(self, priority_override=None):
+        if priority_override is not None:
+            self.PRIORITY = priority_override
 
 
 class Executor(object):
     """
     This class is in charge of all of the threads.  It starts up the threads
-    and cleans up the threads when done.  The two type of threads the
+    and cleans up the threads when finished.  The two type of threads the
     ``Executor``runs is a worker and a print thread.
     """
-    def __init__(self, done, num_threads, result_queue,
-                 quiet, interrupt, max_queue_size, write_queue):
-        self.queue = None
-        self.done = done
+    STANDARD_PRIORITY = 11
+    IMMEDIATE_PRIORITY= 1
+
+    def __init__(self, num_threads, result_queue,
+                 quiet, max_queue_size, write_queue):
+        self._max_queue_size = max_queue_size
+        self.queue = StablePriorityQueue(maxsize=self._max_queue_size,
+                                         max_priority=20)
         self.num_threads = num_threads
         self.result_queue = result_queue
         self.quiet = quiet
-        self.interrupt = interrupt
         self.threads_list = []
-        self._max_queue_size = max_queue_size
         self.write_queue = write_queue
-        self.print_thread = None
-        self.io_thread = None
+        self.print_thread = PrintThread(self.result_queue,
+                                        self.quiet)
+        self.print_thread.daemon = True
+        self.io_thread = IOWriterThread(self.write_queue)
 
     @property
     def num_tasks_failed(self):
@@ -51,17 +63,14 @@ class Executor(object):
         return tasks_failed
 
     def start(self):
-        self.print_thread = PrintThread(self.result_queue, self.done,
-                                        self.quiet, self.interrupt)
-        self.print_thread.daemon = True
-        self.io_thread = IOWriterThread(self.write_queue, self.done)
         self.io_thread.start()
-        self.threads_list.append(self.io_thread)
-        self.queue = NoBlockQueue(self.interrupt, maxsize=self._max_queue_size)
-        self.threads_list.append(self.print_thread)
+        # Note that we're *not* adding the IO thread to the threads_list.
+        # There's a specific shutdown order we need and we're going to be
+        # explicit about it rather than relying on the threads_list order.
+        # See .join() for more info.
         self.print_thread.start()
         for i in range(self.num_threads):
-            worker = Worker(queue=self.queue, done=self.done)
+            worker = Worker(queue=self.queue)
             worker.setDaemon(True)
             self.threads_list.append(worker)
             worker.start()
@@ -73,38 +82,66 @@ class Executor(object):
         LOGGER.debug("Submitting task: %s", task)
         self.queue.put(task)
 
-    def wait(self):
-        """
-        This is the function used to wait on all of the tasks to finish
-        in the ``Executor``.
-        """
-        self.queue.join()
+    def initiate_shutdown(self, priority=STANDARD_PRIORITY):
+        """Instruct all threads to shutdown.
 
-    def join(self):
+        This is a graceful shutdown.  It will wait until all
+        currently queued tasks have been completed before the threads
+        shutdown.  If the task queue is completely full, it may
+        take a while for the threads to shutdown.
+
+        This method does not block.  Once ``initiate_shutdown`` has
+        been called, you can all ``wait_until_shutdown`` to block
+        until the Executor has been shutdown.
+
         """
-        This is used to clean up the ``Executor``.
-        """
-        self.write_queue.put(QUEUE_END_SENTINEL)
-        self.result_queue.put(QUEUE_END_SENTINEL)
+        # Implementation detail:  we only queue the worker threads
+        # to shutdown.  The print/io threads are shutdown in the
+        # ``wait_until_shutdown`` method.
         for i in range(self.num_threads):
-            self.queue.put(QUEUE_END_SENTINEL)
+            LOGGER.debug(
+                "Queueing end sentinel for worker thread (priority: %s)",
+                priority)
+            self.queue.put(ShutdownThreadRequest(priority))
 
+    def wait_until_shutdown(self):
+        """Block until the Executor is fully shutdown.
+
+        This will wait until all worker threads are shutdown, along
+        with any additional helper threads used by the executor.
+
+        """
         for thread in self.threads_list:
-            thread.join()
+            LOGGER.debug("Waiting for thread to shutdown: %s", thread)
+            while True:
+                thread.join(timeout=1)
+                if not thread.is_alive():
+                    break
+            LOGGER.debug("Thread has been shutdown: %s", thread)
+
+        LOGGER.debug("Queueing end sentinel for result thread.")
+        self.result_queue.put(ShutdownThreadRequest())
+        LOGGER.debug("Queueing end sentinel for IO thread.")
+        self.write_queue.put(ShutdownThreadRequest())
+
+        LOGGER.debug("Waiting for result thread to shutdown.")
+        self.print_thread.join()
+        LOGGER.debug("Waiting for IO thread to shutdown.")
+        self.io_thread.join()
+        LOGGER.debug("All threads have been shutdown.")
 
 
 class IOWriterThread(threading.Thread):
-    def __init__(self, queue, done):
+    def __init__(self, queue):
         threading.Thread.__init__(self)
         self.queue = queue
-        self.done = done
         self.fd_descriptor_cache = {}
 
     def run(self):
         while True:
             task = self.queue.get(True)
-            if task is QUEUE_END_SENTINEL:
-                LOGGER.debug("Sentinel received in IO thread, "
+            if isinstance(task, ShutdownThreadRequest):
+                LOGGER.debug("Shutdown request received in IO thread, "
                              "shutting down.")
                 self._cleanup()
                 return
@@ -137,28 +174,26 @@ class Worker(threading.Thread):
     This thread is in charge of performing the tasks provided via
     the main queue ``queue``.
     """
-    def __init__(self, queue, done):
+    def __init__(self, queue):
         threading.Thread.__init__(self)
         # This is the queue where work (tasks) are submitted.
         self.queue = queue
-        self.done = done
 
     def run(self):
         while True:
             try:
                 function = self.queue.get(True)
-                if function is QUEUE_END_SENTINEL:
-                    self.queue.task_done()
+                if isinstance(function, ShutdownThreadRequest):
+                    LOGGER.debug("Shutdown request received in worker thread, "
+                                 "shutting down worker thread.")
                     break
                 try:
+                    LOGGER.debug("Worker thread invoking task: %s", function)
                     function()
                 except Exception as e:
                     LOGGER.debug('Error calling task: %s', e, exc_info=True)
-                self.queue.task_done()
-            except Queue.Empty:
+            except queue.Empty:
                 pass
-            if self.done.isSet():
-                break
 
 
 class PrintThread(threading.Thread):
@@ -181,12 +216,10 @@ class PrintThread(threading.Thread):
             deprecated, will be removed in the future).
 
     """
-    def __init__(self, result_queue, done, quiet, interrupt):
+    def __init__(self, result_queue, quiet):
         threading.Thread.__init__(self)
         self._progress_dict = {}
-        self._interrupt = interrupt
         self._result_queue = result_queue
-        self._done = done
         self._quiet = quiet
         self._progress_length = 0
         self._num_parts = 0
@@ -213,8 +246,11 @@ class PrintThread(threading.Thread):
         while True:
             try:
                 print_task = self._result_queue.get(True)
-                if print_task is QUEUE_END_SENTINEL:
-                    self._result_queue.task_done()
+                if isinstance(print_task, ShutdownThreadRequest):
+                    if self._needs_newline:
+                        sys.stdout.write('\n')
+                    LOGGER.debug("Shutdown request received in print thread, "
+                                 "shutting down print thread.")
                     break
                 LOGGER.debug("Received print task: %s", print_task)
                 try:
@@ -222,18 +258,8 @@ class PrintThread(threading.Thread):
                 except Exception as e:
                     LOGGER.debug("Error processing print task: %s", e,
                                  exc_info=True)
-                finally:
-                    # Because the shutdown logic requires that the print
-                    # queue finish, we need to have all the print tasks
-                    # finished, even if an exception happens trying to print
-                    # them.
-                    self._result_queue.task_done()
-            except Queue.Empty:
+            except queue.Empty:
                 pass
-            if self._done.isSet():
-                if self._needs_newline:
-                    sys.stdout.write('\n')
-                break
 
     def _process_print_task(self, print_task):
         print_str = print_task['message']
@@ -265,7 +291,7 @@ class PrintThread(threading.Thread):
             self._file_count += 1
 
         is_done = self._total_files == self._file_count
-        if not self._interrupt.isSet() and not is_done:
+        if not is_done:
             prog_str = "Completed %s " % self._num_parts
             num_files = self._total_files
             if self._total_files != '...':