+++ tests/unit/customizations/s3/test_executor.py
@@ -10,7 +10,7 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-from tests import unittest
+from tests import unittest, temporary_file
 import os
 import tempfile
 import shutil
@@ -18,7 +18,8 @@ import mock
 from six.moves import queue
 
 from awscli.customizations.s3.executor import IOWriterThread
-from awscli.customizations.s3.executor import QUEUE_END_SENTINEL
+from awscli.customizations.s3.executor import ShutdownThreadRequest
+from awscli.customizations.s3.executor import Executor
 from awscli.customizations.s3.utils import IORequest, IOCloseRequest
 
 
@@ -26,7 +27,7 @@ class TestIOWriterThread(unittest.TestCase):
 
     def setUp(self):
         self.queue = queue.Queue()
-        self.io_thread = IOWriterThread(self.queue, mock.Mock())
+        self.io_thread = IOWriterThread(self.queue)
         self.temp_dir = tempfile.mkdtemp()
         self.filename = os.path.join(self.temp_dir, 'foo')
         # Create the file, since IOWriterThread expects
@@ -39,7 +40,7 @@ class TestIOWriterThread(unittest.TestCase):
     def test_handles_io_request(self):
         self.queue.put(IORequest(self.filename, 0, b'foobar'))
         self.queue.put(IOCloseRequest(self.filename))
-        self.queue.put(QUEUE_END_SENTINEL)
+        self.queue.put(ShutdownThreadRequest())
         self.io_thread.run()
         with open(self.filename, 'rb') as f:
             self.assertEqual(f.read(), b'foobar')
@@ -48,7 +49,7 @@ class TestIOWriterThread(unittest.TestCase):
         self.queue.put(IORequest(self.filename, 6, b'morestuff'))
         self.queue.put(IORequest(self.filename, 0, b'foobar'))
         self.queue.put(IOCloseRequest(self.filename))
-        self.queue.put(QUEUE_END_SENTINEL)
+        self.queue.put(ShutdownThreadRequest())
         self.io_thread.run()
         with open(self.filename, 'rb') as f:
             self.assertEqual(f.read(), b'foobarmorestuff')
@@ -60,10 +61,28 @@ class TestIOWriterThread(unittest.TestCase):
         self.queue.put(IORequest(second_file, 0, b'otherstuff'))
         self.queue.put(IOCloseRequest(second_file))
         self.queue.put(IOCloseRequest(self.filename))
-        self.queue.put(QUEUE_END_SENTINEL)
+        self.queue.put(ShutdownThreadRequest())
 
         self.io_thread.run()
         with open(self.filename, 'rb') as f:
             self.assertEqual(f.read(), b'foobar')
         with open(second_file, 'rb') as f:
             self.assertEqual(f.read(), b'otherstuff')
+
+
+class TestExecutor(unittest.TestCase):
+    def test_shutdown_does_not_hang(self):
+        executor = Executor(2, queue.Queue(), False,
+                            10, queue.Queue(maxsize=1))
+        with temporary_file('rb+') as f:
+            executor.start()
+            class FloodIOQueueTask(object):
+                PRIORITY = 10
+
+                def __call__(self):
+                    for i in range(50):
+                        executor.write_queue.put(IORequest(f.name, 0, b'foobar'))
+            executor.submit(FloodIOQueueTask())
+            executor.initiate_shutdown()
+            executor.wait_until_shutdown()
+            self.assertEqual(open(f.name, 'rb').read(), b'foobar')